Moved global variables to Config.h

+++ src/main.cpp
@@ -71,7 +71,7 @@ int main(int argc, char* argv[]){
 
 	Experiment *experiment;
 
-	unsigned int n_settings = settings.size();
+	long unsigned int n_settings = settings.size();
 
 	for(unsigned int i = 0; i < n_settings; ++i){
